chore: extract zip helper methods into separate class
