+++ node/src/main/java/eu/cloudnetservice/cloudnet/node/command/sub/CommandCluster.java
@@ -27,7 +27,7 @@ import com.google.common.collect.Lists;
 import com.google.common.net.InetAddresses;
 import eu.cloudnetservice.cloudnet.common.column.ColumnFormatter;
 import eu.cloudnetservice.cloudnet.common.column.RowBasedFormatter;
-import eu.cloudnetservice.cloudnet.common.io.FileUtil;
+import eu.cloudnetservice.cloudnet.common.io.ZipUtil;
 import eu.cloudnetservice.cloudnet.common.language.I18n;
 import eu.cloudnetservice.cloudnet.common.log.LogManager;
 import eu.cloudnetservice.cloudnet.common.log.Logger;
@@ -260,7 +260,7 @@ public final class CommandCluster {
     @NonNull String serviceName
   ) {
     // zip the whole directory into a stream
-    var stream = FileUtil.zipToStream(servicePath);
+    var stream = ZipUtil.zipToStream(servicePath);
     // notify the source about the deployment
     source.sendMessage(I18n.trans("command-cluster-push-static-service-starting"));
     // deploy the static service into the cluster
