+++ node/src/main/java/eu/cloudnetservice/cloudnet/node/module/updater/ModuleUpdaterRegistry.java
@@ -35,7 +35,7 @@ public final class ModuleUpdaterRegistry extends DefaultUpdaterRegistry<ModuleUp
     Map<Path, String> moduleNames = new HashMap<>();
     FileUtil.walkFileTree(DefaultModuleProvider.DEFAULT_MODULE_DIR, ($, file) -> {
       // open the jar file
-      FileUtil.openZipFileSystem(file, fs -> {
+      FileUtil.openJarFileSystem(file, fs -> {
         var moduleJson = fs.getPath("module.json");
         if (Files.exists(moduleJson)) {
           // read the file
